Show list of questions and answers
Show the list of questions and answers on the check_answers page.
The list should show the questions in the same order that they appear in
the YAML file.

At the moment there is no guard that checks whether the question was
answered or skipped, but we can add that in later.