Update the feature test
We can't access the session in feature tests, so that means we can't set a value for
`sessions[:live_in_england]` and allow the test to skip straight to the basic needs page.

The proper thing to do would be to add a step for every page, but as we're using this more as
a smoke test, I've changed the the test to start at the beginning of the journey.