Add locale file for find_support form
The aim here is to create a clean YAML file that content designers could edit on github.

To achieve this:
1) Use rails Intl to separate content from code, keeping a single form's
   content in a single yaml file. Stakeholders for this page would be
content designers, we'd want to ensure they are comfortable pulling this
file up on github and making changes to existing content by finding and
replacing values.

2) To support that, I've tried to prevent this YAML file nesting too
   deeply or becoming too long and intimidating. Groups content is
essentially just a title, so that's just an array of key value pairs,
defaults should keep question options dry by defining repeatedly used
keys. Where a special case or override is required special_options: can
be described within a specific question. The intention is to have a view
helper or to research a bit more on intl defaults to ensure we check for
a special_option first and then fall back to defaults, and only error if
there is neither availble and a call is made in a rails view.

3) content != config. Locale files are just to provide content, there
   should be no indication of question structure from this file (next
commit will have more on that). So it should be noted:
- The existance of a question locale does NOT mean the question is
  included in a form,
- The existance of an option does NOT mean it is included in a question
- The order of the locale file does NOT determine the order of the form
- Groups are only included here to describe their headers, NOT which
  question is included in a group