Ensure user has answered the first question
The user must answer the first question before they can answer any
others.  If a user is deep-linked into the form, they will be redirected
to the first question.

Using a helper method to determine this (for simplicity in testing),
since we cannot modify the session store in most request tests.