Return leave home to the block list
Leave home currently does not have any results, as such, it doesn't have a representation in the YAML file, because there's nothing to put there. This causes an error where the locale really helpfully makes a placeholder object that then causes this code to fireball horribly. It'd be good to add a test for this usescase. The solution should be, either allow the code to recognise these placeholders and filter them, or just add leave_home to the en.yml with a placeholder datastructure that has no links