Expire a user's session if they visit an invalid question
For questions after the "what do you need help with" question, users
will get an `undefined method `+' for nil:NilClass` if they visit a
question which is not one they should be seeing (based on their answers
to the "what do you need help with question").

This expires their session (and redirects to the appropriate information
page) in the event that they visit an invalid question (e.g. if they use
a bookmark from a previous journey).