Don't show users previous answers from session
For the find coronavirus support service, we don't have a
check-your-answers page, so having people be able to go back to the
start of their journey and see their previous answers has limited value.

There is a risk that another user of the same computer could start a new
journey before a session expires, and see answers from a previous user's
session. In some situations this could be harmful.

This commit updates the views so they don't show people's answers, but
leaves all other session handling in place (so the results page still
works in exactly the same way).

Note that the back button in browsers is "clever" with this, so
previous form responses may still be present when clicking back, but
should not be present when hard refreshing the page or navigating
directly to the URL.

To reproduce the old "bad" behaviour:

* Go through the journey as far as https://find-coronavirus-support.service.gov.uk/need-help-with
* Click Continue
* Close the tab
* Open https://find-coronavirus-support.service.gov.uk/need-help-with in a new tab
* Observ that your selections are still there