Add route for exporting usage data
This endpoint allows us to access user data without needing a developer
to run a rake task.

At the moment, this exports CSV but can easily be changed to JSON by
changing the response type and not calling the `produce_csv` method.

The ActiveRecord query looks complex, but we are utilising Postgres's
JSONB query and counts to avoid the need for the Rails application to
retrieve all the data then do the sorting (this is all handled by
Postgres and we just receive the count values).