Update data export tests and delimiter
Changes the delimiter in the csv file from `,` to `|`. This is because
many of the title and option label strings already contain commas, so
the CSV generator was wrapping this strings (and only these strings) in
double quotes and the tests were compensating for this but also wrapping
those strings. But if the text ever changes and the comma is removed,
the tests would still need to be updated, which is the situation we are
trying to avoid.

Also updates all of tests to get their string values from the I18n file.