Bugfix: I'm not sure should redirect users to first question
When a user has all questions to answer (eg when they answer
'I'm not sure' to the question what help do you need),
we should set their questions_to_answer to be all_questions.

Currently we set the questions_to_answer to be empty, which
means users are redirected to a nil path next_quest(nil).

This wasn't working because all_questions became an array of
symbols, not strings.

Resolves https://sentry.io/organizations/govuk/issues/1621299615
https://trello.com/c/Pw8wJYuW/288