Fix bug with Timecop
Freezing time was causing feature tests to fail, we suspect this is due
to session data being out of date.  Un-freezeing time at the end of each
test in the AbleToLeaveController means that test time now matches system
time, and the tests pass.

This only appears to effect the full test suite and not when running just
feature tests.  Not 100% sure at this point why this is true.

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>
Co-authored-by: Jonathan Hallam <jonathan.hallam@digital.cabinet-office.gov.uk>
Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>