Fix test cases for results helper
The new method we wrote broke a few tests for the results helper. This is because it was consuming i18n and filtering it down in code in a way that wasn't happening before. This commit tries to decouple our testing from our copy. The solution is far from perfect and I'd like to see these tests and the code that they test refactored in the future to make testing easier.