Change label for "Feeling unsafe" on filter question
We currently use the group labels as checkbox options in the "What do
you need help" filter question.  For "Feeling unsafe" we want to have a
different label for the checkbox in this question, but retain the
existing group title.

Therefore adding a `need_help_with_label` value to the group.  This will
also us to override the label that is displayed in the "What do you need
help with" question for a specific group.