Fix jasmine tests for enhanced ecommerce
The tests for the result page tracking checked that the `ga()` function
**wasn't** called when a user hadn't opted in to tracking. This was done by
spying on `ga` and seeing if this had been called.

If the tests ran in order, that `ga()` function would exist. If they didn't
run in order, the function wouldn't exist - and errors would be thrown...

Because we can't test that the `ga()` function works correctly in sending data
to Google Analytics, we can happily just create an empty function to mock
`ga()` so we can test that the function is called with the correct attributes.