Change data export dates to be optional
Due to the way the where was setup, just substiutted the params for a default of '2020-01-01' (i.e. before this service existed) and an end date of '3000-01-1' (i.e. all time, I really hope this is all over by then..)

The aim of this is to give an effectively all time export unless dates are explicitly provided

On second thought instead of hard coding a date into the far future,
lets do today + 1 day (in seconds)

On third thought, lets use time.zone