Simplify actions file further
This adds the JS tests to be part of the default rake task so that CI
steps can be replicated with a single command.

It then makes a number of simplifications to the actions file in line
with https://github.com/alphagov/govuk-developer-docs/pull/2533, mostly
evident in the simpler DB connection and single environment variable
definitions.