Add support and advice to content exporter
A refactor to expand the exporter to deal with new results section
Support and Advice. These sections will be siblings of items and always
appear below them.

They are optional and may not exist, but will always exist where there
are items.

The exporter can now handle these allowing us to populate the content
spreadsheet.