Conditional Support + Advice Section in results page. https://trello.com/c/RWmAHtdt/400-results-page-update-2-3-conditional-support-advice-section Add markup with a condition for "Support and Advice" sections within a group.
For a given group, if there are "Support and Advice" links, then render the heading + any links. If there are no links then do not show the heading.

[See this card where CSS work is also underway](https://trello.com/c/Am4icV9a/145-results-page-update-change-heading-sizes), this card can focus on the markup and leave the css be done there.

NB: A "group" may not have any Support + Advice links. But if there are Support and advice links there will always be "main" links directly under the `h3`