Fix date range selector in data export
The tests were only checking for what is there, but not what isn’t
there, so it didn’t check that a date that should be filtered out isn’t
being returned anymore.

It was the guard around the `where` clause that was causing the problem.
So it has been split out of the main query.

Thanks to @Rosa-Fox for spotting the bug and writing the test to catch
it.