Move the able to leave question
This moves the "are you able to leave your house" question and makes sure it's only asked if the user has indicated that they're worried about getting food. This touches a lot of the application because previously that question was being used to store responses, a functionality that now has to be given to all of the controllers.

This also now checks for the existance of session data by checking that the user has answered the first question. This is because the last question is no longer asked 100% of the time. It also means that the method checking for the last question being answered can be removed.