Fix flakey smoke test
The smoke tests were failing as a result of the javascript not being
able to execute in time. 'page.body' as introduced in this
commit(https://github.com/alphagov/govuk-coronavirus-find-support
/pull/138/commits/9466d919d14d4027ff22be2719ed480cb699f02b)
returns a string while 'page' will wait for the contents of the page to load.
This means the content of the page cannot be checked asynchronously.