Rename the living with vulnerable route
A previous commit changed the copy for the "Are you worried about going into work" question. It didn't change the route so that was still `/living-withvulnerable`, a phrase that wasn't in the copy any more. This commit changes the names of the controller, view and spec for that path, changes the routes and adds a redirect for the old route. This means that the route the user sees now matches the copy on the page.