Rake task to check problematic/broken links in our content.
Yaml files are searched for all the links (internal gov.uk and external links)
The links are grouped into a set for each file, this allows us to combine the sets and automatically remove duplicates.

Each link is then passed into the link checker api (GdsApi::LinkCheckerApi) to check if there are issues such as broken, redirecting, slow or malicious links.

Future enhancements are to handle internal gov.uk links separately and pass into publishing -api to check for unpublished or draft items.

Co-authored-by: Dan Williams <dan@ithium.net>
Co-authored-by: Steve Holmes <steholmes@gmail.com>
Co-authored-by: Rob Nicols <git@nicholshayes.co.uk>