Remove the google sheet importer.
##What
We want to change the way we import content into the results page, from a Google Sheets to uploading directly onto the YAML file.

Why
The Google Sheets method was built in case people didn't know how to use YAML. But it's a bit fiddly to work with and the team know how to use YAML, so let's make it easier for the team.

Agreed to keep the Google sheet exporter, so that Content Designers have a cleaner way of viewing the results. Remove only the importer for now.