Move "Are you off work ill" question
Moved the "Are you off work ill?" question to always be asked after the
"Are you worried about going into work?" question instead of being asked
in the unemployed section. This and future commits are apart of work to
make to clearier how to get advice about if you are off work ill.